# app-int/src/main/resources/openapi/specs/public.yml
openapi: 3.0.3

info:
  title: eForm validation service REST-API
  description: >
    Über diesen Endpunkt können Bekanntmachungen im eForms Format validiert werden.

  contact:
    name: oeffentliche-vergabe-support@nortal.com
    email: oeffentliche-vergabe-support@nortal.com

  version: 1.0.1

  tags:
    - name: Validation Endpoint
    - name: Business Document Validation Endpoint

paths:
  /v1/validation:
    post:
      tags:
        - Validation Endpoint
      operationId: validate
      summary: Eine Bekanntmachung validieren
      description: >
        Über diesen Endpunkt können Bekanntmachungen im eForms Format validiert werden. Die 
        Anfrage gibt das Ergebnis der Validierung inklusive möglicher Fehler und Warnungen zurück.
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - eformsVersion
                - sdkType
                - eforms
              type: object
              required:
                - eformsVersion
                - sdkType
                - eforms
              properties:
                eformsVersion:
                  pattern: "^[\\d]\\.[\\d]$"
                  type: string
                  example: '1.0'
                sdkType:
                  pattern: "^eforms-(eu|de|EU|DE)$"
                  type: string
                  example: eforms-eu
                eforms:
                  format: binary
                  description: Eine Bekanntmachung im eForms Format
                  type: string
                xsdValidation:
                  type: boolean
                  required: false
                  description: XSD Validierung Flag (true, wenn weggelassen)
                schematronValidation:
                  type: boolean
                  required: false
                  description: Schematron Validierung Flag (true, wenn weggelassen)
            encoding:
              eforms:
                contentType: application/octet-stream
      responses:
        "200":
          description: >
            OK
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Validation"
        "400":
          description: >
            BAD REQUEST
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/400ApiError"
        "415":
          description: >
            NO VALID CONTENT TYPE
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/415ApiError"
        "500":
          description: >
            INTERNAL SERVER ERROR
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/500ApiError"
  /v1/validation-business-document:
    post:
      tags:
        - Business Document Validation Endpoint
      operationId: validateBusinessDocument
      summary: Ein business_document.xml validieren
      description: >
        Über diesen Endpunkt kann ein business_document.xml validiert werden. 
        Die Anfrage gibt das Ergebnis der Validierung inklusive möglicher Fehler und Warnungen zurück.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - business_document
              properties:
                business_document:
                  type: string
                  format: binary
                  required: true
                  description: Ein business_document.xml
      responses:
        "200":
          description: >
            OK
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/BusinessDocumentValidation"
        "400":
          description: >
            BAD REQUEST
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/400ApiError"
        "405":
          description: >
            METHOD NOT ALLOWED
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/405ApiError"
        "415":
          description: >
            NO VALID CONTENT TYPE
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/415ApiError"
        "500":
          description: >
            INTERNAL SERVER ERROR
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/500ApiError"

components:
  schemas:
    Validation:
      xml:
        name: "validation"
      type: object
      properties:
        warnings:
          type: array
          items:
            "$ref": "#/components/schemas/ValidationWarning"
        errors:
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
        validatedEformsVersion:
          type: string
          example: eforms-sdk-1.0.0
        valid:
          type: boolean
          example: true
    BusinessDocumentValidation:
      xml:
        name: "business_document_validation"
      type: object
      properties:
        warnings:
          type: array
          items:
            "$ref": "#/components/schemas/ValidationWarning"
        errors:
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
        valid:
          type: boolean
          example: true
    400ApiError:
      xml:
        name: "400ApiError"
      type: object
      properties:
        reason:
          type: string
          example: MALFORMED_XML
        description:
          type: string
          example: Die Anfrage wurde vom Server zurückgewiesen, da die gesendeten Daten im Request-Body nicht akzeptiert wurden.
    405ApiError:
      xml:
        name: "405ApiError"
      type: object
      properties:
        reason:
          type: string
          example: METHOD_NOT_ALLOWED
        description:
          type: string
          example: HTTP-Methode ist nicht erlaubt.
    415ApiError:
      xml:
        name: "415ApiError"
      type: object
      properties:
        reason:
          type: string
          example: NO_VALID_CONTENT_TYPE
        description:
          type: string
          example: Die zugelieferten Bekanntmachungsdaten wurden in einem nicht akzeptablen Format gesendet. Nur `text/xml` wird akzeptiert
    500ApiError:
      xml:
        name: "500ApiError"
      type: object
      properties:
        reason:
          type: string
          example: UNKNOWN_ERROR
        description:
          type: string
          example: Wenden Sie sich bitte an unser Support-Team.
    ValidationWarning:
      xml:
        name: "validationWarning"
      type: object
      properties:
        rule:
          type: string
          example: rule|text|BR-OPT-00301-1180
        ruleContent:
          type: string
          example: count(cac:MediationParty/cac:PartyIdentification/cbc:ID) > 0
        description:
          type: string
          example: OPT-301-Lot-Mediator is mandatory for a notice with subtype 16
        path:
          type: string
          example:  /ContractNotice/cac:ProcurementProjectLot/cac:TenderingTerms/cac:AppealTerms
        type:
          type: string
          example: SCHEMATRON
    ValidationError:
      xml:
        name: "validationError"
      type: object
      properties:
        rule:
          type: string
          example: rule|text|BR-BT-00747-0052
        ruleContent:
          type: string
          example: normalize-space(.) = ('ef-stand', 'other', 'sui-act', 'tp-abil')
        description:
          type: string
          example: BT-747-Lot must contain one of the values of the code list selection-criterion
        path:
          type: string
          example: /ContractNotice/cac:ProcurementProjectLot/cac:TenderingTerms/ext:UBLExtensions/ext:UBLExtension/ext:ExtensionContent/efext:EformsExtension/efac:SelectionCriteria/cbc:CriterionTypeCode
        type:
          type: string
          enum:
            - XSD
            - SCHEMATRON
          example: SCHEMATRON